Added few changes to testbench and a new file to help organize the code
