+++ system.mk
@@ -89,7 +89,7 @@ PYTHON_DIR:=$(SW_DIR)/python
 
 #hw paths
 HW_DIR=$(ROOT_DIR)/hardware
-SIM_DIR=$(HW_DIR)/simulation/$(SIMULATOR)
+SIM_DIR=$(HW_DIR)/simulation/cocotb
 ASIC_DIR=$(HW_DIR)/asic
 BOARD_DIR ?=$(shell find hardware -name $(BOARD))
